requirements.txt
# For rendering
glcontext
moderngl
PyOpenGL

# Others
jupyter
pillow
opencv-python
trimesh
torch
matplotlib
numpy